Move SlackClient to Common
