+++ kifi-backend/modules/common/app/com/keepit/slack/SlackClientModule.scala
@@ -3,7 +3,6 @@ package com.keepit.slack
 import com.google.inject.{ Provides, Singleton }
 import com.keepit.common.net.HttpClient
 import net.codingwell.scalaguice.ScalaModule
-import play.api.Mode.Mode
 
 import scala.concurrent.ExecutionContext
 
@@ -16,15 +15,5 @@ case class ProdSlackClientModule() extends SlackClientModule {
   @Provides
   def slackClient(httpClient: HttpClient, ec: ExecutionContext): SlackClient = {
     new SlackClientImpl(httpClient, ec)
-  }
-}
-
-case class FakeSlackClientModule() extends SlackClientModule {
-  def configure() {}
-
-  @Singleton
-  @Provides
-  def slackClient(): SlackClient = {
-    new FakeSlackClientImpl()
   }
 }
