+++ kifi-backend/modules/common/app/com/keepit/slack/models/SlackAPIFailure.scala
